+++ testsuite/tests/deriving/should_compile/T9968a.hs
@@ -5,4 +5,4 @@ module T9968 where
 import Data.Bifunctor
 
 data Blah a b = A a | B b
-  deriving (Bifunctor)
+  deriving (Functor, Bifunctor)